# asg_form/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ASG 赛事官网-后端API文档",
    "description": "这是由罗澜使用ASP.NET.Core开发的ASG赛事组后端系统，包括官网和后台管理系统。使用 sqlserver作为数据库，identity框架进行账号控制。",
    "version": "V 1.9.6"
  },
  "paths": {
...
    "/api/v1/enroll": {
      "post": {
        "tags": [
          "login"
        ],
        "summary": "注册",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "captoken",
            "description": "谷歌人机验证token",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "用户信息",
            "schema": {
              "$ref": "#/definitions/newuser_get"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/newuser_get"
            }
          }
        }
      }
    },
...
    "/api/v1/login": {
      "post": {
        "tags": [
          "login"
        ],
        "summary": "登陆",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "用户信息",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/newuser_get"
            }
          }
        }
      }
    },
...
    "/api/v1/okemail": {
      "post": {
        "tags": [
          "login"
        ],
        "summary": "确认邮件验证码",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "邮箱",
            "schema": {
              "$ref": "#/definitions/ok_email_get"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/newuser_get"
            }
          }
        }
      }
    },
...
    "/api/v1/user/name": {
      "post": {
        "tags": [
          "login"
        ],
        "summary": "修改中文名称",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "newchinaname",
            "description": "新的中文名称",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "LoginRequest": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "userName": {
          "type": "string"
        },
        "normalizedUserName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "normalizedEmail": {
          "type": "string"
        },
        "emailConfirmed": {
          "type": "boolean"
        },
        "passwordHash": {
          "type": "string"
        },
        "securityStamp": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "phoneNumberConfirmed": {
          "type": "boolean"
        },
        "twoFactorEnabled": {
          "type": "boolean"
        },
        "lockoutEnd": {
          "format": "date-time",
          "type": "string"
        },
        "lockoutEnabled": {
          "type": "boolean"
        },
        "accessFailedCount": {
          "format": "int32",
          "type": "integer"
        },
        "creationTime": {
          "format": "date-time",
          "type": "string"
        },
        "haveform": {
          "$ref": "#/definitions/form"
        },
        "isbooking": {
          "type": "boolean"
        },
        "chinaname": {
          "type": "string"
        },
        "officium": {
          "type": "string"
        }
      }
    },
...
    "form": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "piaoshu": {
          "format": "int32",
          "type": "integer"
        },
        "time": {
          "format": "date-time",
          "type": "string"
        },
        "team_name": {
          "type": "string"
        },
        "team_password": {
          "type": "string"
        },
        "team_tel": {
          "type": "string"
        },
        "logo_uri": {
          "type": "string"
        },
        "role": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/role"
          }
        }
      }
    },
...
    "newuser_get": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "chinaname": {
          "type": "string"
        },
        "eMail": {
          "type": "string"
        }
      }
    },
    "ok_email_get": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
...
    "role": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "form": {
          "$ref": "#/definitions/form"
        },
        "role_id": {
          "type": "string"
        },
        "role_name": {
          "type": "string"
        },
        "role_lin": {
          "type": "string"
        }
      }
    },
...
  }
}